Update Math.js to fix Issue 256
The fix for this issue is the same as the fix for Issue 244: if the results are out of range, try again.  The function was restructured to invoke `internalRandomGenerator` in one location after the function's parameters' values are determined.  Handling the degenerate case where `a===b` and maintaining a not-so-infinite retry loop remain intact.

The header comment was cleaned up.  Ill chosen words like "low" and "high" were removed.  Similarly, the function's arguments are now assigned to `a` and `b`.